+++ src/main/java/com/caved_in/commons/block/BlockHitAction.java
@@ -1,7 +1,8 @@
 package com.caved_in.commons.block;
 
 import org.bukkit.block.Block;
+import org.bukkit.entity.Player;
 
 public interface BlockHitAction {
-	public void onHit(Block block);
+	public void onHit(Player player, Block block);
 }