+++ src/main/java/com/caved_in/commons/entity/CreatureHitAction.java
@@ -1,7 +1,8 @@
 package com.caved_in.commons.entity;
 
 import org.bukkit.entity.LivingEntity;
+import org.bukkit.entity.Player;
 
 public interface CreatureHitAction {
-	public void onHit(LivingEntity entity);
+	public void onHit(Player player, LivingEntity entity);
 }